♻️ Reduce the number of lines in the src/index.ts file
Move getSassOptions() and validateSassOptions() functions definitions to src/options/sass.ts file.
